[REFACTOR] : ViewHolder 분리
